separate Header in components, fix warning with key
