+++ src/components/DepartmentInfo/DepartmentInfo.js
@@ -1,11 +1,12 @@
 import React, {Component} from 'react'
 import {Ingredients} from '../Ingredients/Ingredients'
+import shortid from 'shortid'
 
 class DepartmentInfo extends Component{
 
   render(){
     return(
-      <li key={this.props.key}>
+      <li key={shortid.generate()}>
         <h2>{this.props.department}</h2>
         <Ingredients ingredients = {this.props.ingredients} />
       </li>
