+++ src/components/Header/Header.test.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './Header';
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.unmountComponentAtNode(div);
+});
